Fix #69: Do not unwrap when extracting data from responses (#72)
There still are a bunch of other unwraps in the code (some of which you
can even see in this patch), mostly parsing either CBOR or other binary
stuff in the crypto-section.

Not sure, if we want to address them here as well, and if so, how. 
For CBOR-related things, we could basically also just return the serde
errors?