Implement basic extensions infrastructure (#77)
A few things going on here:
- Previously, `AuthenticatorData` was just a ByteBuf. I added it as a
struct with de/serialize-functions.
- Input and Output of the extensions vary for given commands, so I made
`AuthenticatorData` generic over that, and specify which
extension-struct it will contain inside `MakeCredential` or
`GetAssertion`. This should be totally invisible to the end-user, who
will just get 'the right thing'. It does, however, make the
de/serialize-code a bit messy.
- `hmac-secret` extension needs more work, but this PR was already
getting quite big, so I figured, this would be better in a follow-up PR.